backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
qwen-tts
torch
torchaudio
soundfile
requests
